Move CallableActivation to interfaces
