+++ neuro_san/internals/interfaces/callable_activation.py
